Update `rcrc` with change from thoughtbot/dotfiles
I noticed that after pulling down the latest from thoughtbot/dotfiles
and running `rcup`, cloning of Git repositories was broken.

In general, if one is maintaining their own `rcrc` file in a local
dotfiles repo, for maximum parity with thoughtbot/dotfiles, it's
necessary to keep the local version in sync.

The fix also necessitated deleting the previously, unintentionally,
symlinked `~/.git_template/HEAD`, and re-running `rcup`.

The commit in question was
https://github.com/thoughtbot/dotfiles/commit/56d614f8067c275f4d7b53f3f3a30a5effbeded9